+++ src/modules/getConfig/index.ts
@@ -1,7 +1,7 @@
 import { Prefix } from "../../shared/const/Prefix"
 import { getConfigRules } from "../getConfigRules"
 
-import type { MasterRecord } from "../../shared/types"
+import type { MasterRecord } from "../../shared/types/MasterRecord"
 import type { Options } from "../../shared/types/Options"
 import type { RuleName } from "../../shared/types/RuleName"
