+++ src/shared/types/EsLintRuleBase/index.ts
@@ -1,6 +1,5 @@
 import type { Options } from "../Options"
 import type { TSESLint } from "@typescript-eslint/utils"
 
-
 type MessageId = string
 export type EsLintRuleBase = TSESLint.RuleModule<MessageId, Options>
