+++ src/shared/types/MasterRecord/index.ts
@@ -1,0 +1,5 @@
+import type { EsLintRuleBase } from "../EsLintRuleBase"
+import type { Options } from "../Options"
+import type { RuleName } from "../RuleName"
+
+export type MasterRecord = [RuleName, EsLintRuleBase, Options]
